Unify podcast-row component; apply Theme typography
Three near-duplicate podcast-row layouts (PodcastsView's private PodcastRow,
ProfileView.subscriptionsSection ad-hoc HStack, SearchView's
PodcastSearchRow) collapse into one shared `PodcastRow(podcast:density:)`
with two densities:

- .regular — 80pt artwork, 2-line description, inline SubscribeButton
- .compact — 56pt artwork, single-line name + artist meta

ProfileView's subscriptions list and SearchView's podcasts results both
switch to the compact density. SubscribeButton itself moves to its own
file so any caller can drop it in.

EpisodeRow and PodcastDetailView's hero name now consume the Theme
typography tokens (`itemTitle`, `screenTitle`) instead of literal fonts.
