+++ LiboLibo/Features/Podcasts/PodcastsView.swift
@@ -104,76 +104,6 @@ private struct PodcastListItem: View {
     }
 }
 
-/// Типографика как в `EpisodeRow` (Фид):
-/// .headline — название, .subheadline — превью описания. Внизу — кнопка
-/// подписки в формате иконки.
-private struct PodcastRow: View {
-    let podcast: Podcast
-
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            AsyncImage(url: podcast.artworkUrl) { phase in
-                switch phase {
-                case .success(let image):
-                    image.resizable().aspectRatio(contentMode: .fill)
-                default:
-                    Color.secondary.opacity(0.15)
-                }
-            }
-            .frame(width: 80, height: 80)
-            .clipShape(RoundedRectangle(cornerRadius: Theme.radiusMedium))
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(podcast.name)
-                    .font(.headline)
-                    .foregroundStyle(.primary)
-
-                if let desc = podcast.description {
-                    let preview = desc.firstSentences(maxCount: 2)
-                    if !preview.isEmpty {
-                        Text(preview)
-                            .font(.subheadline)
-                            .foregroundStyle(.secondary)
-                            .padding(.top, 2)
-                    }
-                }
-
-                HStack(spacing: 8) {
-                    SubscribeButton(podcast: podcast)
-                }
-                .padding(.top, 4)
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(.vertical, 8)
-        .contentShape(Rectangle())
-    }
-}
-
-/// Иконка-toggle подписки. Логика как у `DownloadButton`: подписан —
-/// заполненная галочка в брендовом цвете, не подписан — контурный плюс
-/// в `.secondary`.
-private struct SubscribeButton: View {
-    let podcast: Podcast
-    @Environment(SubscriptionsService.self) private var subscriptions
-
-    var body: some View {
-        Button {
-            subscriptions.toggle(podcast)
-        } label: {
-            Image(systemName: subscriptions.isSubscribed(podcast)
-                  ? "checkmark.circle.fill"
-                  : "plus.circle")
-                .font(.title3)
-                .foregroundStyle(subscriptions.isSubscribed(podcast) ? Color.liboRed : .secondary)
-                .frame(width: 28, height: 28, alignment: .leading)
-                .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel(subscriptions.isSubscribed(podcast) ? "Отписаться" : "Подписаться")
-    }
-}
-
 #Preview {
     PodcastsView()
         .environment(PodcastsRepository())
